feat: navermap directions, geocode 적용, deploy.sh update

+++ src/main/java/com/rissins/manna/dto/response/LocationResponse.java
@@ -6,7 +6,7 @@ import lombok.Data;
 @Data
 @Builder
 public class LocationResponse {
-    int lat;
-    int lon;
-    int middlePoint;
+    double lat;
+    double lon;
+    double middlePoint;
 }
